correct error

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,3 @@
-#include <stdlib.h>
-
 #include "function_pointers.h"
 
 /**
@@ -8,10 +6,12 @@
  * @array: array of elements
  * @size: size of the array
  * @action: pointer to the function you need to use
+ *
+ * Return: Nothing
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	if (!array || !action)
 		return;
